pyproject.toml
[project]
name = "thinkonward-heat-signature-zero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "ipykernel",
    "plotly",
    "pandas",
    "nbformat",
    "joblib",
    "jax>=0.8.1",
    "jaxlib>=0.8.1",
    "jax-cuda12-pjrt==0.8.1; sys_platform == 'linux'",
    "jax-cuda12-plugin==0.8.1; sys_platform == 'linux'",
    "nvidia-cublas-cu12==12.9.1.4; sys_platform == 'linux'",
    "nvidia-cuda-cupti-cu12==12.9.79; sys_platform == 'linux'",
    "nvidia-cuda-nvcc-cu12==12.9.86; sys_platform == 'linux'",
    "nvidia-cuda-nvrtc-cu12==12.9.86; sys_platform == 'linux'",
    "nvidia-cuda-runtime-cu12==12.9.79; sys_platform == 'linux'",
    "nvidia-cudnn-cu12==9.17.1.4; sys_platform == 'linux'",
    "nvidia-cufft-cu12==11.4.1.4; sys_platform == 'linux'",
    "nvidia-cusolver-cu12==11.7.5.82; sys_platform == 'linux'",
    "nvidia-cusparse-cu12==12.5.10.65; sys_platform == 'linux'",
    "nvidia-nccl-cu12==2.28.9; sys_platform == 'linux'",
    "nvidia-nvjitlink-cu12==12.9.86; sys_platform == 'linux'",
    "nvidia-nvshmem-cu12==3.4.5; sys_platform == 'linux'",
    "ipywidgets>=8.1.8",
    "mlflow>=2.9.0",
    "pyyaml>=6.0",
    "cma>=3.3.0",
]
